fix: UPDATE_TRANSACTIONS reducer should not overwrite previously saved transactions
